Stock maintenace added to menu and data type for price was changed to decimal

+++ Isaris.DataAccess/CotizacionDAL.cs
@@ -66,7 +66,7 @@ namespace Isaris.DataAccess
 
             }
         }
-        public static void UpdateTotal(int idInvoice, double total)
+        public static void UpdateTotal(int idInvoice, decimal total)
         {
             using (MySqlConnection conn = new MySqlConnection(ConfigurationManager.ConnectionStrings["default"].ToString()))
             {
